+++ Assets/VRGameCamera.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82e92378d2cfbb3488ab362933775863
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
