+++ Assets/VRGameCamera/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 873c2f996f454e74b910be9335df5bff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
