+++ Assets/VRGameCamera/Prefabs/VRGameCamera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fe278dcd0d36b25419057bc434448d24
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
