+++ Assets/VRGameCamera/Prefabs/VRGameCameraUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 91a38336d3734cc4499df2e3fde87fc2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
