+++ Assets/VRGameCamera/RecordTexture.renderTexture
