+++ Assets/VRGameCamera/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5079a611ba07e845a6df441bc71fb4e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
